rlm-core/pyproject.toml
[project]
name = "rlm-core"
version = "0.1.0"
description = "Local RLM Engine for WorkGraph OS - Cost-optimized verification with local SLMs"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "instructor>=0.6.0",
    "httpx>=0.26.0",
    "ollama>=0.1.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    "python-jose[cryptography]>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rlm_core"]
